% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3_metadata.R
\name{Gen3MetadataParser}
\alias{Gen3MetadataParser}
\title{Create a Gen3 metadata parser object}
\usage{
Gen3MetadataParser(key_file)
}
\arguments{
\item{key_file}{Character string path to the JSON key file containing Gen3 credentials}
}
\value{
A gen3_metadata object with credentials and base URL configured
}
\description{
This function creates a new Gen3 metadata parser object by loading
credentials from a key file. Use this object to interact with the Gen3 API.
}
