% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-gen3_metadata.R
\name{print}
\alias{print}
\alias{print.gen3_metadata}
\title{Print gen3_metadata object}
\usage{
print(gen3_metadata)

\method{print}{gen3_metadata}(gen3_metadata)
}
\arguments{
\item{gen3_metadata}{A gen3_metadata object}
}
\value{
The gen3_metadata object (invisibly)

The gen3_metadata object (invisibly)
}
\description{
Generic function to print a summary of a gen3_metadata object.

This method prints a formatted summary of a gen3_metadata object.
}
