% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gen3_metadata.R
\name{print.gen3_metadata}
\alias{print.gen3_metadata}
\title{Print gen3_metadata object summary}
\usage{
\method{print}{gen3_metadata}(gen3_metadata)
}
\arguments{
\item{gen3_metadata}{A gen3_metadata object}
}
\value{
The gen3_metadata object (invisibly)
}
\description{
This method prints a formatted summary of a gen3_metadata object.
}
